refactor: support image in svg and prefetch external lins in element
