+++ src/satori.ts
@@ -9,6 +9,7 @@ import svg from './builder/svg.js'
 import { segment } from './utils.js'
 import { detectLanguageCode, LangCode, Locale } from './language.js'
 import getTw from './handler/tailwind.js'
+import { preProcessNode } from './handler/preprocess.js'
 
 // We don't need to initialize the opentype instances every time.
 const fontCache = new WeakMap()
@@ -86,6 +87,8 @@ export default async function satori(
   // the Chinese as missing font because we have removed the Chinese as a fallback.
   // To address this situation, we may need to add `processedWordsMissingFont`
   const processedWordsMissingFonts = new Set()
+
+  await preProcessNode(element)
 
   const handler = layout(element, {
     id: 'id',
